+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add compatibility with browserpass via `--pass-compat=browserpass`
 - Add compatibility mode `username` for a `username:` prefix
 - Add `--pass-cmd` to allow specifying pass's location or script name.
+- Using `--pass-prefix=''` prevents creation of a prefix: `web/address/...` becomes `address/...`
 
 ##### 0.7.0
 - Fix PK11 slot memory leak